+++ excel/src/main/java/br/com/fag/domain/usecases/MenorValorParaApostasComCincoDezenas.java
@@ -7,7 +7,7 @@ import br.com.fag.domain.entities.Resultado;
 public class MenorValorParaApostasComCincoDezenas implements IAnalise {
   @Override
   public void run(Aposta aposta) {
-    Double valor = aposta.getRateios()[1].getValorRateio();
+    Long valor = aposta.getRateios()[1].getValorRateio();
     if(valor < Resultado.menorValorParaApostasComCincoDezenas) {
       Resultado.menorValorParaApostasComCincoDezenas = valor;
     }
