+++ excel/src/main/java/br/com/fag/domain/usecases/MenorValorParaApostasComQuatroDezenas.java
@@ -7,7 +7,7 @@ import br.com.fag.domain.entities.Resultado;
 public class MenorValorParaApostasComQuatroDezenas implements IAnalise {
   @Override
   public void run(Aposta aposta) {
-    Double valor = aposta.getRateios()[2].getValorRateio();
+    Long valor = aposta.getRateios()[2].getValorRateio();
     if(valor < Resultado.menorValorParaApostasComQuatroDezenas) {
       Resultado.menorValorParaApostasComQuatroDezenas = valor;
     }
