+++ excel/src/main/java/br/com/fag/domain/usecases/MenorValorParaApostasComSeisDezenas.java
@@ -7,7 +7,7 @@ import br.com.fag.domain.entities.Resultado;
 public class MenorValorParaApostasComSeisDezenas implements IAnalise {
   @Override
   public void run(Aposta aposta) {
-    Double valor = aposta.getRateios()[0].getValorRateio();
+    Long valor = aposta.getRateios()[0].getValorRateio();
     if(valor < Resultado.menorValorParaApostasComSeisDezenas) {
       Resultado.menorValorParaApostasComSeisDezenas = valor;
     }
